+++ bahz_testing.R
@@ -11,277 +11,83 @@ library(rstan)
 options(mc.cores = parallel::detectCores())
 library(tidyverse)
 library(loo)
-# Getting the minimal pre-compiled model to run -------------------
+
+
+
+# Testing geno models -----------------------------------------------------
 # Generate a dataset
 set.seed(22)
-data <- sim_geno_cline(transect_distances = seq(-300,300,20), n_ind = 40, Fis = 0,
-                       decrease = F, center = 10, width = 35, pmin = 0.08, pmax = .95)
+data2 <- sim_geno_cline(transect_distances = seq(-300,300,20), n_ind = 40, Fis = .9,
+                       decrease = F, center = 10, width = 35, pmin = 0.08, pmax = .95, deltaR = 12, tauR = 0.25)
 
 plot(x = data$transectDist, y = data$emp.p)
 lines(x = data$transectDist, y = data$cline.p)
+
+
+library(bahz)
 
 # Fit the model
 
-geno_fit_bi <- fit_geno_cline(data = data, prior_file = "~/Desktop/geno_priors.yaml",
+none_bi <- fit_geno_cline2(data = data2, prior_file = "~/Desktop/geno_priors.yaml",
                            type = "bi", tails = "none")
-geno_fit_multi <- fit_geno_cline(data = data, prior_file = "~/Desktop/geno_priors.yaml",
-                           type = "multi", tails = "none")
+none_multi <- fit_geno_cline2(data = data2, prior_file = "~/Desktop/geno_priors.yaml",
+                              type = "multi", tails = "none")
 
-cline_summary(geno_fit_bi)
-
-stanfit <- geno_fit_bi
-
-
-plot(z$transectDist, z$p_mean, type = "l")
-lines(z$transectDist, z$p_median, col = "red")
-
-plot_cline(stanfit = geno_fit_bi, data = data, add.obs = T, confidence = T, method = "HPDI",
-           prob = 0.99)
-
-plot_geno_cline(geno_fit_bi, data = data, add.obs.freqs = T, confidence = T, main = "Binomial")
-plot_cline(geno_fit_bi, data = data, add.obs = F, confidence = F, main = "Binomial_C")
-
-
-plot_geno_cline(geno_fit_multi, data = data, add.obs.freqs = T, confidence = T, main = "Multinomial")
-plot_cline(geno_fit_multi, data = data, add.obs = T, confidence = T, main = "Multinomial_C")
-
-
-# And a phenotypic model
-x <- seq(-100, 100, 20)
-pheno <- sim_pheno_cline(transect_distances = x, n_ind = as.integer(abs(rnorm(length(x), mean = 15, sd = 9))),
-                         sigma = rnorm(length(x), mean = 2, sd = 0.75),
-                         decrease = F, center = 15, width = 30, pmin = 12, pmax = 24)
-site.means <- pheno %>%
-  group_by(transectDist) %>%
-  summarize(mean.pheno = mean(traitValue))
-plot(pheno$transectDist, pheno$traitValue)
-lines(x, site.means$mean.pheno, col = "red")
-
-constant_sig <- fit_pheno_cline(data = pheno,
-                                prior_file = "~/Desktop/pheno_priors.yaml", pheno_variance = "constant",
-                                chains = 4)
-pool_sig <- fit_pheno_cline(data = pheno,
-                            prior_file = "~/Desktop/pheno_priors.yaml", pheno_variance = "pooled",
-                            chains = 4)
-ind_sig <- fit_pheno_cline(data = pheno,
-                           prior_file = "~/Desktop/pheno_priors.yaml", pheno_variance = "independent",
-                           chains = 4)
-
-plot_pheno_cline(constant_sig, data = pheno, add.obs.pheno = T, confidence = T, main = "constant")
-plot_cline(constant_sig, data = pheno, add.obs = T, confidence = T, main = "constant_C")
-plot_pheno_cline(pool_sig, data = pheno, add.obs.pheno = T, confidence = T, main = "pooled")
-plot_cline(pool_sig, data = pheno, add.obs = T, confidence = T, main = "pooled_C")
-plot_pheno_cline(ind_sig, data = pheno, add.obs.pheno = T, confidence = T, main = "independent", prob = 0.5)
-plot_cline(ind_sig, data = pheno, add.obs = T, confidence = T, main = "independent_C", prob = 0.5)
-
-
-load("tests/testthat/ref_simulated_pheno_clines.Rda")
-load("tests/testthat/ref_geno_stanfit.Rda")
-ref_geno_stanfit <- ref_stanfit
-save(ref_geno_stanfit, file = "tests/testthat/ref_geno_stanfit.Rda")
-
-
-
-x <- seq(-75, 75, 25)
-ref.pheno.data <- sim_pheno_cline(transect_distances = x, n_ind = 10,
-                         sigma = 1,
-                         decrease = F, center = 15, width = 30, pmin = 12, pmax = 24)
-
-
-ref_pheno_stanfit <- fit_pheno_cline(data = ref.pheno.data, prior_file = "~/Desktop/pheno_priors.yaml")
-save(ref_pheno_stanfit, file = "tests/testthat/ref_pheno_stanfit.Rda")
-save(ref.pheno.data, file = "tests/testthat/ref_pheno_data.Rda")
-
-# NEED TO GENERATE A NEW DATASET FOR THE PLOT CLINE TESTS FOR PHENO, THE A>REF AND B>REF ETC FILES ALL MAKE GIGANTC STANFITS.
-
-# Fit the model, binomial
-fit_none_b <- fit_geno_cline(data = data, prior_file = "all_betas.yaml",
-                           type = "bi", tails = "none")
-fit_left_b <- fit_geno_cline(data = data, prior_file = "all_betas.yaml",
-                      type = "bi", tails = "left")
-fit_right_b <- fit_geno_cline(data = data, prior_file = "all_betas.yaml",
-                           type = "bi", tails = "right")
-fit_mirror_b <- fit_geno_cline(data = data, prior_file = "all_betas.yaml",
-                           type = "bi", tails = "mirror")
-fit_ind_b <- fit_geno_cline(data = data, prior_file = "all_betas.yaml",
-                            type = "bi", tails = "ind")
-
-plot_geno_cline(fit_none_b, data = data, add.obs.freqs = T, col = "red")
-plot_geno_cline(fit_none_b, data = data, add.obs.freqs = T, col = "xxx")
-plot_geno_cline(fit_none_b, data = dplyr::slice(data, -1), add.obs.freqs = T, point.col = "red")
-
-
-plot_geno_cline(fit_left_b, data = data, add.obs.freqs = T, col = "red")
-plot_geno_cline(fit_right_b, data = data, add.obs.freqs = T, col = "red")
-plot_geno_cline(fit_mirror_b, data = data, add.obs.freqs = T, col = "red")
-plot_geno_cline(fit_ind_b, data = data, add.obs.freqs = T, col = "red")
-
-plot_geno_cline(fit_none_b, data = data, main = "test", col = "red", xlab = "distance", ylab = "allele frequency")
-
-general_cline_eqn(c(8,10), decrease = F, center = 5, width = 8)
-
-cline_summary(fit_none_b)
-cline_summary(fit_left_b)
-cline_summary(fit_right_b)
-cline_summary(fit_mirror_b)
-cline_summary(fit_ind_b)
-
-z <- fit_geno_cline(data = data, prior_file = "prior_config_test1.yaml", type = "bi", tails = "none", chains = 1)
-
-
-
-inits <- prep_init_list("prior_config_test1.yaml", tails = "none", chains = as.integer(1), type = "geno")
-priors <- prep_prior_list("prior_config_test1.yaml")
-
-z@inits
-
-
-plot(predict_geno_cline(fit_none_b, data = data)$transectDist, predict_geno_cline(fit_none_b, data = data)$p, type = "l")
-
-
-lines(-300:300, plot_cline(fit_left_b)$p, type = "l", col = "red")
-lines(-300:300, plot_cline(fit_right_b)$p, type = "l", col = "blue")
-lines(-300:300, plot_cline(fit_mirror_b)$p, type = "l", col = "orange")
-lines(-300:300, plot_cline(fit_ind_b)$p, type = "l", col = "green")
-
-# Fit the model, binomial
-fit_none_m <- fit_geno_cline(data = data, prior_file = "prior_config_template.yaml",
-                           type = "bi", tails = "none")
-fit_left_m <- fit_geno_cline(data = data, prior_file = "prior_config_template.yaml",
+left_bi <- fit_geno_cline2(data = data, prior_file = "~/Desktop/geno_priors.yaml",
+                               type = "bi", tails = "left")
+left_multi <- fit_geno_cline2(data = data, prior_file = "~/Desktop/geno_priors.yaml",
                            type = "multi", tails = "left")
-fit_right_m <- fit_geno_cline(data = data, prior_file = "prior_config_template.yaml",
+
+right_bi <- fit_geno_cline2(data = data, prior_file = "~/Desktop/geno_priors.yaml",
+                           type = "bi", tails = "right")
+right_multi <- fit_geno_cline2(data = data, prior_file = "~/Desktop/geno_priors.yaml",
                             type = "multi", tails = "right")
-fit_mirror_m <- fit_geno_cline(data = data, prior_file = "prior_config_template.yaml",
+
+mirror_bi <- fit_geno_cline2(data = data, prior_file = "~/Desktop/geno_priors.yaml",
+                            type = "bi", tails = "mirror")
+mirror_multi <- fit_geno_cline2(data = data, prior_file = "~/Desktop/geno_priors.yaml",
                              type = "multi", tails = "mirror")
-fit_ind_m <- fit_geno_cline(data = data, prior_file = "prior_config_template.yaml",
+
+ind_bi <- fit_geno_cline2(data = data, prior_file = "~/Desktop/geno_priors.yaml",
+                             type = "bi", tails = "ind")
+ind_multi <- fit_geno_cline2(data = data, prior_file = "~/Desktop/geno_priors.yaml",
                           type = "multi", tails = "ind")
 
-z <- predict_cline(stanfit = fit_none_m, distance = 0:500, confidence = T, prob = 0.94)
-z2 <- predict_cline(stanfit = fit_left_m, distance = 0:500, confidence = T, prob = 0.95)
-z3 <- predict_cline(stanfit = fit_right_m, distance = 0:500, confidence = T, prob = 0.95)
-z4 <- predict_cline(stanfit = fit_mirror_m, distance = 0:500, confidence = T, prob = 0.95)
-z5 <- predict_cline(stanfit = fit_ind_m, distance = 0:500, confidence = T, prob = 0.95)
 
 
-install.packages("memoise")
-
-library(memoise)
-
-mem_pred_cline <- memoise(predict_cline)
-
-test <- memoise(function(x) {x + 1})
-
-test(1)
-
-z <- predict_cline(stanfit = fit_none_m, distance = -300:300, confidence = T, prob = 0.95)
-z2 <- predict_cline(stanfit = fit_none_m, distance = -300:300, confidence = T, prob = 0.95)
-
-data$transectDist[2] <- -281
-
-plot_geno_cline(stanfit = fit_none_m, data = data, add.obs.freqs = T, confidence = T)
-plot_geno_cline(stanfit = fit_none_m, data = data, add.obs.freqs = T, confidence = T,
-                cline.col = "orange", point.col = "yellow", prob = 0.5,
-                lwd = 4, pch = 22, cex = 4, bg = "green")
-
-par()
-
-ggplot() +
-  geom_ribbon(fill = "grey90",
-              aes(x = transectDist, ymin = low_0.95_HPDI, ymax = up_0.95_HPDI),
-              data = z) +
-  geom_point(aes(x = transectDist, y = emp.p), data = data) +
-  geom_line(aes(x = transectDist, y = p), data = z) +
-  ggtitle("none")
-
-ggplot() +
-  geom_ribbon(fill = "grey90",
-              aes(x = transectDist, ymin = low_0.95_HPDI, ymax = up_0.95_HPDI),
-              data = z2) +
-  geom_point(aes(x = transectDist, y = emp.p), data = data) +
-  geom_line(aes(x = transectDist, y = p), data = z2) +
-  ggtitle("left")
-
-ggplot() +
-  geom_ribbon(fill = "grey90",
-              aes(x = transectDist, ymin = low_0.95_HPDI, ymax = up_0.95_HPDI),
-              data = z3) +
-  geom_point(aes(x = transectDist, y = emp.p), data = data) +
-  geom_line(aes(x = transectDist, y = p), data = z3) +
-  ggtitle("right")
-
-ggplot() +
-  geom_ribbon(fill = "grey90",
-              aes(x = transectDist, ymin = low_0.95_HPDI, ymax = up_0.95_HPDI),
-              data = z4) +
-  geom_point(aes(x = transectDist, y = emp.p), data = data) +
-  geom_line(aes(x = transectDist, y = p), data = z4) +
-  ggtitle("mirror")
-
-ggplot() +
-  geom_ribbon(fill = "grey90",
-              aes(x = transectDist, ymin = low_0.95_HPDI, ymax = up_0.95_HPDI),
-              data = z5) +
-  geom_point(aes(x = transectDist, y = emp.p), data = data) +
-  geom_line(aes(x = transectDist, y = p), data = z5) +
-  ggtitle("ind")
-
-rethinking::compare(fit)
-
-cline_summary(fit_left_m)
-
-prep_init_list("prior_config_template.yaml", tails = "ind", chains = as.integer(1))
-
-fit_none_m@inits
-fit_left_m@inits
-
-cline_summary(fit_none_m)
-cline_summary(fit_left_m)
-cline_summary(fit_right_m)
-cline_summary(fit_mirror_m)
-cline_summary(fit_ind_m)
+rethinking::compare(none_bi, left_bi, right_bi, mirror_bi, ind_bi)
+cline_summary(none_bi)
+cline_summary(none_bi, show.all = T)
+cline_summary(none_multi)
+cline_summary(left_bi)
+cline_summary(left_multi)
+cline_summary(right_bi)
+cline_summary(right_multi)
+cline_summary(mirror_bi)
+cline_summary(mirror_multi)
+cline_summary(ind_bi)
+cline_summary(ind_multi)
 
 
-xs <- seq(-300,300,20)
-ys <- rep(NA, length(xs))
-for (i in 1:length(xs)) {
-  ys[i] <- general_cline_eqn(transectDist = xs[i], decrease = T,
-                         center = 7.49, width = 40.20, pmin = 0.03, pmax = 0.96,
-                        deltaL = 0.2, tauL = 0.65, deltaR = 0.2, tauR = 0.65)
-}
-plot(x = data$transectDist, y = data$emp.p)
-lines(x = data$transectDist, y = data$cline.p)
-lines(x = xs, y = ys, col = "red")
 
-plot(z1)
+# Testing pheno models ----------------------------------------------------
 
-z1 <- loo::loo(fit_none_m, r_eff = relative_eff(fit_none_m))
-z2 <- loo::loo(fit_left_m, r_eff = relative_eff(fit_left_m))
-z3 <- loo::loo(fit_right_m, r_eff = relative_eff(fit_right_m))
-z4 <- loo::loo(fit_mirror_m, r_eff = relative_eff(fit_mirror_m))
-z5 <- loo::loo(fit_ind_m, r_eff = relative_eff(fit_ind_m))
+set.seed(839)
+data <- sim_pheno_cline(transect_distances = seq(-200, 150, length.out = 12), n_ind = as.integer(rnorm(n = 12, mean = 30, sd = 7)),
+                        sigma = abs(rnorm(n = 12, mean = 40, sd = 8)), decrease = T, center = 9, width = 72, pmin = 220, pmax = 762)
 
-rethinking::compare(fit_none, fit_left, fit_right, fit_mirror, fit_ind)
-loo::compare(z1, z2, z3, z4, z5)
+constant <- fit_pheno_cline(data, prior_file = "~/Desktop/pheno_priors.yaml", pheno_variance = "constant")
+independent <- fit_pheno_cline(data, prior_file = "~/Desktop/pheno_priors.yaml", pheno_variance =  "independent")
+pooled <- fit_pheno_cline(data, prior_file = "~/Desktop/pheno_priors.yaml", pheno_variance = "pooled")
 
-# testing the new width check in prep_init_list
-i <- 8000
-z <- prep_init_list("prior_config_template.yaml", tails = "right", chains = as.integer(i))
-widths <- rep(as.numeric(NA), times = i)
-for (width in 1:i) {
-  widths[width] <- z[[width]]$width
-}
-unique(unlist(widths) < 0) # Never makes a value below 0.
+names(constant)
+cline_summary(constant, show.all = F)
+cline_summary(constant, show.all = T)
 
-test <- function(...) {
-  stan.args <- names(sapply(match.call(), deparse))[-1]
-  if (c("control") %in% stan.args == F) {
-    result <- list(adapt_delta = 0.95)
-  } else{
-    result <- which(stan.args == "control")
-  }
-}
+cline_summary(independent, show.all = F)
+cline_summary(pooled)
 
-
+rm(list = ls())
 # Will stan run with no priors? -------------------------------------------
 
 # Yes, stan will run with no priors! I think the default, then, is that it