+++ R/fit_geno_cline.R
@@ -29,6 +29,7 @@
 #' E.g., specifying a different number of total iterations will reset the
 #' adapt_delta to the \code{stan} default of 0.8.
 #'
+#'
 #' @param data A dataframe containing your cline data. See
 #'   \code{\link{prep_geno_data}} for possible formats.
 #' @param prior_file The path to the \code{.yaml} file which contains the
@@ -44,6 +45,10 @@
 #'   initialization values for \code{stan}, to be used instead of the
 #'   \code{bahz} default of random initialization values from the prior. See
 #'   \code{\link[rstan]{stan}} for details on how to specify the init list.
+#' @param ignore_data Used for doing prior predictive checks. If FALSE, the default,
+#'   the model is fit normally. If TRUE, the data are ignored during model fitting.
+#'   This is useful for double-checking the prior distributions or parameters, and
+#'   for doing prior predictive checks of the model.
 #' @param ... Arguments to be passed to \code{stan}, e.g., number of iterations,
 #'   warmup period, etc. See \code{\link[rstan]{sampling}} and the
 #'   \code{control} argument in \code{\link[rstan]{stan}} for information on
@@ -84,11 +89,12 @@
 fit_geno_cline <- function(data, prior_file,
                       type = c("bi", "multi"),
                       tails = c("none", "left", "right", "mirror", "ind"),
-                      chains = 4, init = NULL, ...) {
+                      chains = 4, ignore_data = FALSE, init = NULL, ...) {
   # Argument checking
   type <- match.arg(type, several.ok = F)
   tails <- match.arg(tails, several.ok = F)
-  assertthat::assert_that(is.numeric(chains) == T, msg = "chains must be numeric")
+  assertthat::assert_that(is.numeric(chains), msg = "chains must be numeric")
+  assertthat::assert_that(is.logical(ignore_data), msg = "ignore_data must be either TRUE or FALSE")
   ch <- as.integer(chains)
 
   # Calling internal functions
@@ -104,6 +110,13 @@ fit_geno_cline <- function(data, prior_file,
     init_list <- init
   }
 
+  # Process ignore data
+  if (ignore_data) {# if ignoring the data
+    ignoreData <- list(ignoreData = as.integer(1))
+  } else {# otherwise, use the data
+    ignoreData <- list(ignoreData = as.integer(0))
+  }
+
 
   # Find location of the model in the stanmodels object that matches
   # the desired model provide by the user
@@ -111,10 +124,10 @@ fit_geno_cline <- function(data, prior_file,
 
   # Pass everything to stan
   if (length(eval(substitute(alist(...)))) > 0) {# if user supplies extra parameters to go to Stan
-    clinefit <- rstan::sampling(object = stanmodels[[model_index]], data = c(stan_data, prior_list),
+    clinefit <- rstan::sampling(object = stanmodels[[model_index]], data = c(stan_data, prior_list, ignoreData),
                                 chains = ch, init = init_list, ...)
   } else {# otherwise, use the bahz defaults
-    clinefit <- rstan::sampling(object = stanmodels[[model_index]], data = c(stan_data, prior_list),
+    clinefit <- rstan::sampling(object = stanmodels[[model_index]], data = c(stan_data, prior_list, ignoreData),
                                 chains = ch, init = init_list, control = list(adapt_delta = 0.95))
     }