+++ bahz_testing.R
@@ -17,41 +17,38 @@ library(loo)
 # Testing geno models -----------------------------------------------------
 # Generate a dataset
 set.seed(22)
-data2 <- sim_geno_cline(transect_distances = seq(-300,300,20), n_ind = 40, Fis = .9,
+data <- sim_geno_cline(transect_distances = seq(-300,300,20), n_ind = 40, Fis = .9,
                        decrease = F, center = 10, width = 35, pmin = 0.08, pmax = .95, deltaR = 12, tauR = 0.25)
 
 plot(x = data$transectDist, y = data$emp.p)
 lines(x = data$transectDist, y = data$cline.p)
 
 
-library(bahz)
 
-# Fit the model
+none_bi <- fit_geno_cline(data = data2, prior_file = "~/Desktop/geno_priors.yaml",
+                              type = "bi", tails = "none", ignore_data = F)
+none_multi <- fit_geno_cline(data = data2, prior_file = "~/Desktop/geno_priors.yaml",
+                              type = "multi", tails = "none", ignore_data = F)
 
-none_bi <- fit_geno_cline2(data = data2, prior_file = "~/Desktop/geno_priors.yaml",
-                           type = "bi", tails = "none")
-none_multi <- fit_geno_cline2(data = data2, prior_file = "~/Desktop/geno_priors.yaml",
-                              type = "multi", tails = "none")
+left_bi <- fit_geno_cline(data = data, prior_file = "~/Desktop/geno_priors.yaml",
+                               type = "bi", tails = "left", ignore_data = F)
+left_multi <- fit_geno_cline(data = data, prior_file = "~/Desktop/geno_priors.yaml",
+                           type = "multi", tails = "left", ignore_data = F)
 
-left_bi <- fit_geno_cline2(data = data, prior_file = "~/Desktop/geno_priors.yaml",
-                               type = "bi", tails = "left")
-left_multi <- fit_geno_cline2(data = data, prior_file = "~/Desktop/geno_priors.yaml",
-                           type = "multi", tails = "left")
+right_bi <- fit_geno_cline(data = data, prior_file = "~/Desktop/geno_priors.yaml",
+                           type = "bi", tails = "right", ignore_data = F)
+right_multi <- fit_geno_cline(data = data, prior_file = "~/Desktop/geno_priors.yaml",
+                            type = "multi", tails = "right", ignore_data = F)
 
-right_bi <- fit_geno_cline2(data = data, prior_file = "~/Desktop/geno_priors.yaml",
-                           type = "bi", tails = "right")
-right_multi <- fit_geno_cline2(data = data, prior_file = "~/Desktop/geno_priors.yaml",
-                            type = "multi", tails = "right")
+mirror_bi <- fit_geno_cline(data = data, prior_file = "~/Desktop/geno_priors.yaml",
+                            type = "bi", tails = "mirror", ignore_data = F)
+mirror_multi <- fit_geno_cline(data = data, prior_file = "~/Desktop/geno_priors.yaml",
+                             type = "multi", tails = "mirror", ignore_data = F)
 
-mirror_bi <- fit_geno_cline2(data = data, prior_file = "~/Desktop/geno_priors.yaml",
-                            type = "bi", tails = "mirror")
-mirror_multi <- fit_geno_cline2(data = data, prior_file = "~/Desktop/geno_priors.yaml",
-                             type = "multi", tails = "mirror")
-
-ind_bi <- fit_geno_cline2(data = data, prior_file = "~/Desktop/geno_priors.yaml",
-                             type = "bi", tails = "ind")
-ind_multi <- fit_geno_cline2(data = data, prior_file = "~/Desktop/geno_priors.yaml",
-                          type = "multi", tails = "ind")
+ind_bi <- fit_geno_cline(data = data, prior_file = "~/Desktop/geno_priors.yaml",
+                             type = "bi", tails = "ind", ignore_data = F)
+ind_multi <- fit_geno_cline(data = data, prior_file = "~/Desktop/geno_priors.yaml",
+                          type = "multi", tails = "ind", ignore_data = F)
 
 
 
@@ -68,6 +65,16 @@ cline_summary(mirror_multi)
 cline_summary(ind_bi)
 cline_summary(ind_multi)
 
+# Prior predictive checks
+library(bayesplot)
+ppc
+
+
+y_rep_bi <- as.matrix(none_bi, pars = names(none_bi)[which(str_detect(names(none_bi), pattern = "^y_rep"))])
+y_rep_multi <- as.matrix(none_multi, pars = names(none_multi)[which(str_detect(names(none_multi), pattern = "^y_rep"))])
+
+
+ppc_intervals(y = data$AA*2 + data$Aa, yrep = y_rep_bi, prob = 0.99)
 
 
 # Testing pheno models ----------------------------------------------------
@@ -76,9 +83,9 @@ set.seed(839)
 data <- sim_pheno_cline(transect_distances = seq(-200, 150, length.out = 12), n_ind = as.integer(rnorm(n = 12, mean = 30, sd = 7)),
                         sigma = abs(rnorm(n = 12, mean = 40, sd = 8)), decrease = T, center = 9, width = 72, pmin = 220, pmax = 762)
 
-constant <- fit_pheno_cline(data, prior_file = "~/Desktop/pheno_priors.yaml", pheno_variance = "constant")
-independent <- fit_pheno_cline(data, prior_file = "~/Desktop/pheno_priors.yaml", pheno_variance =  "independent")
-pooled <- fit_pheno_cline(data, prior_file = "~/Desktop/pheno_priors.yaml", pheno_variance = "pooled")
+constant <- fit_pheno_cline(data, prior_file = "~/Desktop/pheno_priors.yaml", pheno_variance = "constant", ignore_data = F)
+independent <- fit_pheno_cline(data, prior_file = "~/Desktop/pheno_priors.yaml", pheno_variance =  "independent", ignore_data = F)
+pooled <- fit_pheno_cline(data, prior_file = "~/Desktop/pheno_priors.yaml", pheno_variance = "pooled", ignore_data = F)
 
 names(constant)
 cline_summary(constant, show.all = F)
@@ -87,7 +94,37 @@ cline_summary(constant, show.all = T)
 cline_summary(independent, show.all = F)
 cline_summary(pooled)
 
-rm(list = ls())
+
+
+# Some notes on prior predictive checks -----------------------------------
+# Note that for parameters with limits (like width), values
+# that fall outside the allowed threshold are ignored.
+
+# So, if the prior for center is normal with mean 50 and sd 50, the
+# actual prior distribution in the end is:
+
+z <- rnorm(n = 1000, mean = 50, sd = 50)
+mean(z[z>0]) # 61, not
+mean(z) # ~50
+
+# To get bulk and tail ESS ------------------------------------------------
+# A little odd and annoying that these functions don't just work on
+# stanfit objects.
+# But, easiest thing to do is use the monitor function in rstan.
+
+# Can run the function on the stanfit object by itself,
+# Which apparently isn't supposed to work.
+monitor(none_bi)
+
+# This sort of works, but gets the number of warmup iterations wrong also:
+monitor(as.array(none_bi), warmup = 0)
+
+# This method seems to work properly:
+monitor(extract(none_bi, permuted = F, inc_warmup = T))
+
+
+# Commented on a Github issue about this.
+
 # Will stan run with no priors? -------------------------------------------
 
 # Yes, stan will run with no priors! I think the default, then, is that it